Extract more logic into submodule from execution::context
